fix output data type of image classification

+++ src/transformers/pipelines/image_classification.py
@@ -23,6 +23,7 @@ if is_tf_available():
     from ..models.auto.modeling_tf_auto import TF_MODEL_FOR_IMAGE_CLASSIFICATION_MAPPING_NAMES
 
 if is_torch_available():
+    import torch
     from ..models.auto.modeling_auto import MODEL_FOR_IMAGE_CLASSIFICATION_MAPPING_NAMES
 
 logger = logging.get_logger(__name__)
@@ -179,7 +180,7 @@ class ImageClassificationPipeline(Pipeline):
         if top_k > self.model.config.num_labels:
             top_k = self.model.config.num_labels
 
-        outputs = model_outputs["logits"][0]
+        outputs = model_outputs["logits"][0].to(torch.float32)
         outputs = outputs.numpy()
 
         if function_to_apply == ClassificationFunction.SIGMOID:
